feat: Add Core Data to persist pins
In the Core Data model a pin entity exists, to save the longitude and latitude values
of created pins. When the app starts and the map view appears, the saved pins will
be added to the map.

Whenever a new pin is dropped or deleted, the pin will be added or deleted from the
database.